experiments cleanup
