+++ go/cs/reservationstore/db_manip_test.go
@@ -21,9 +21,14 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	base "github.com/scionproto/scion/go/cs/reservation"
 	"github.com/scionproto/scion/go/cs/reservation/e2e"
 	"github.com/scionproto/scion/go/cs/reservation/segment"
+	"github.com/scionproto/scion/go/cs/reservation/segment/admission"
+	stateless "github.com/scionproto/scion/go/cs/reservation/segment/admission/impl"
 	"github.com/scionproto/scion/go/cs/reservation/segmenttest"
+	"github.com/scionproto/scion/go/cs/reservation/sqlite"
+	"github.com/scionproto/scion/go/cs/reservation/test"
 	"github.com/scionproto/scion/go/cs/reservationstorage/backend"
 	"github.com/scionproto/scion/go/lib/colibri/reservation"
 	"github.com/scionproto/scion/go/lib/util"
@@ -38,7 +43,6 @@ func AddSegmentReservation(t testing.TB, db backend.DB, ASID string, count int) 
 	r := newTestSegmentReservation(t, ASID) // the suffix will be overwritten
 	for i := 0; i < count; i++ {
 		r.Path = segmenttest.NewPathFromComponents(0, "1-"+ASID, i, 1, "1-ff00:0:2", 0)
-		// r.Indices = segment.Indices{}
 		err := db.NewSegmentRsv(ctx, r)
 		require.NoError(t, err, "iteration i = %d", i)
 	}
@@ -64,12 +68,24 @@ func AddE2EReservation(t testing.TB, db backend.DB, ASID string, count int) {
 	}
 }
 
+type testCapacities struct {
+	Cap    uint64
+	Ifaces []uint16
+}
+
+var _ base.Capacities = (*testCapacities)(nil)
+
+func (c *testCapacities) IngressInterfaces() []uint16           { return c.Ifaces }
+func (c *testCapacities) EgressInterfaces() []uint16            { return c.Ifaces }
+func (c *testCapacities) Capacity(from, to uint16) uint64       { return c.Cap }
+func (c *testCapacities) CapacityIngress(ingress uint16) uint64 { return c.Cap }
+func (c *testCapacities) CapacityEgress(egress uint16) uint64   { return c.Cap }
+
 // newTestSegmentReservation creates a segment reservation
 func newTestSegmentReservation(t testing.TB, ASID string) *segment.Reservation {
 	t.Helper()
 	r := segment.NewReservation()
 	r.Path = segment.ReservationTransparentPath{}
-	// r.ID.ASID = xtest.MustParseAS("ff00:0:1")
 	r.ID.ASID = xtest.MustParseAS(ASID)
 	r.Ingress = 0
 	r.Egress = 1
@@ -95,8 +111,6 @@ func newTestE2EReservation(t testing.TB, ASID string) *e2e.Reservation {
 		ID: *e2eIDFromRaw(t, ASID, "00000000000000000001"),
 		SegmentReservations: []*segment.Reservation{
 			newTestSegmentReservation(t, ASID),
-			// newTestSegmentReservation(t), // TODO change 2nd and 3rd seg rsvs.
-			// newTestSegmentReservation(t),
 		},
 	}
 	_, err := rsv.NewIndex(util.SecsToTime(1))
@@ -128,4 +142,81 @@ func e2eIDFromRaw(t testing.TB, ASID, suffix string) *reservation.E2EID {
 	ID, err := reservation.NewE2EID(xtest.MustParseAS(ASID), xtest.MustParseHexString(suffix))
 	require.NoError(t, err)
 	return ID
+}
+
+func newDB(t testing.TB) backend.DB {
+	t.Helper()
+	db, err := sqlite.New("file::memory:")
+	require.NoError(t, err)
+	// db.SetMaxOpenConns(10)
+	return db
+}
+
+func newCapacities() base.Capacities {
+	return &testCapacities{
+		Cap:    1024 * 1024, // 1GBps
+		Ifaces: []uint16{1, 2},
+	}
+}
+
+func newAdmitter(cap base.Capacities) admission.Admitter {
+	admitter := &stateless.StatelessAdmission{
+		Capacities: cap,
+		Delta:      1,
+	}
+	return admitter
+}
+
+// newTestRequest creates a request ID ff00:1:1 beefcafe
+func newTestSegmentRequest(t testing.TB, ASID string, ingress, egress uint16,
+	minBW, maxBW reservation.BWCls) *segment.SetupReq {
+
+	t.Helper()
+
+	ID := segmentIDFromRaw(t, ASID, "beefcafe")
+	path := test.NewTestPath()
+	meta, err := base.NewRequestMetadata(path)
+	require.NoError(t, err)
+	return &segment.SetupReq{
+		Request: segment.Request{
+			RequestMetadata: *meta,
+			ID:              *ID,
+			Timestamp:       util.SecsToTime(1),
+			Ingress:         ingress,
+			Egress:          egress,
+		},
+		MinBW:     minBW,
+		MaxBW:     maxBW,
+		SplitCls:  2,
+		PathProps: reservation.StartLocal | reservation.EndLocal,
+	}
+}
+
+func newTestE2ESetupRequest(t testing.TB, ASID string) *e2e.SetupReq {
+	t.Helper()
+
+	ID := e2eIDFromRaw(t, ASID, "beefcafebeefcafebeef")
+	path := test.NewTestPath()
+	baseReq, err := e2e.NewRequest(util.SecsToTime(1), ID, 1, path)
+	require.NoError(t, err)
+	segmentRsvs := []reservation.SegmentID{
+		*segmentIDFromRaw(t, ASID, "00000001"),
+		*segmentIDFromRaw(t, "ff00:2:2", "beefcafe"),
+		*segmentIDFromRaw(t, "ff00:3:3", "beefcafe"),
+	}
+	ASCountPerSegment := []uint8{4, 4, 5}
+	trail := []reservation.BWCls{5, 5}
+	setup, err := e2e.NewSetupRequest(baseReq, segmentRsvs, ASCountPerSegment, 5, trail)
+	require.NoError(t, err)
+	return setup
+}
+
+func newTestE2ESuccessReq(t testing.TB, ASID string) *e2e.SetupReqSuccess {
+	token, err := reservation.TokenFromRaw(
+		xtest.MustParseHexString("16ebdb4f0d042500003f001002bad1ce003f001002facade"))
+	require.NoError(t, err)
+	return &e2e.SetupReqSuccess{
+		SetupReq: *newTestE2ESetupRequest(t, ASID),
+		Token:    *token,
+	}
 }
